// src/app/models/User.ts
export default class User {
    id: string = "";
    token: string = '';
    firstName: string = '';
    lastName: string = '';
    username: string = '';
    password: string = '';
    passwordCheck?: string = '';
    ID: number | string = '';
    city: string = '';
    address: string = '';
    role: number = 0;
}